+++ src/main/java/ru/knastnt/meshtest/repository/CrudProfileRepository.java
@@ -3,7 +3,7 @@ package ru.knastnt.meshtest.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.knastnt.meshtest.model.Profile;
 
-public interface CrudProfileRepository extends JpaRepository<Profile, Long> {
+public interface CrudProfileRepository extends JpaRepository<Profile, Integer> {
 
     Profile getByEmailIgnoreCase(String email);
